#290 Add Breadcrumbs to settings page
