+++ client/app/routes/dashboard/settings/organization/members.tsx
@@ -5,12 +5,9 @@ import { useOrganizationContext } from "~/contexts/OrganizationContext";
 import Skeleton from "~/components/atoms/skeleton/Skeleton";
 import { useForm } from "react-hook-form";
 import Button from "~/components/atoms/button/Button";
-import {
-  Dialog,
-  DialogTrigger,
-  DialogContent,
-} from "~/components/atoms/dialog/Dialog";
+import { Dialog, DialogContent } from "~/components/atoms/dialog/Dialog";
 import { AvatarIcon } from "~/components/atoms/avatar/Avatar";
+import Breadcrumbs from "~/components/organisms/breadcrumbs/Breadcrumbs";
 
 interface FormInput {
   email: string;
@@ -54,64 +51,15 @@ export default function Members() {
   }
 
   return (
-    <div className="flex flex-col px-8 py-4">
-      <div className="flex min-h-10 w-full items-center justify-between">
-        <h1 className="text-xl font-bold">Members</h1>
-        {organization.isOwner && (
-          <Dialog
-            open={showAddMemberDialog}
-            onOpenChange={setShowAddMemberDialog}
-          >
-            <DialogTrigger>
-              <Button className="w-32">Invite Member</Button>
-            </DialogTrigger>
-            <DialogContent>
-              <div className="flex flex-col gap-4">
-                <div className="flex flex-col gap-2">
-                  <h1>Invite Member</h1>
-                  <p className="text-mauve-11 text-sm">
-                    Invite members via email. They'll receive a link to join
-                    your organization.
-                  </p>
-                </div>
-                <form
-                  className="flex flex-col gap-3"
-                  onSubmit={handleSubmit(onInviteMember)}
-                >
-                  <input
-                    type="email"
-                    className="border-mauve-6 text-mauve-12 placeholder:text-mauve-11 bg-gray-2 w-full rounded-md border p-2 text-sm"
-                    placeholder="Email*"
-                    {...register("email")}
-                  />
-                  <div className="flex justify-end gap-2">
-                    <Button
-                      type="button"
-                      intent="secondary"
-                      className="w-24"
-                      onClick={() => {
-                        setShowAddMemberDialog(false);
-                        reset();
-                        sendInviteMutation.reset();
-                      }}
-                    >
-                      Cancel
-                    </Button>
-                    <Button
-                      className="h-10 w-24"
-                      type="submit"
-                      disabled={!emailInput || sendInviteMutation.isPending}
-                    >
-                      Invite
-                    </Button>
-                  </div>
-                </form>
-              </div>
-            </DialogContent>
-          </Dialog>
-        )}
-      </div>
-      <div className="w-full pt-10.5">
+    <>
+      <Breadcrumbs
+        crumbs={[
+          { label: "Settings" },
+          { label: "Organization" },
+          { label: "Members" },
+        ]}
+      />
+      <div className="flex flex-col px-4 py-4">
         <div className="border-mauve-6 overflow-hidden rounded-md border text-sm shadow-xs">
           <table className="w-full border-collapse">
             <thead className="h-14">
@@ -124,7 +72,6 @@ export default function Members() {
                 </th>
               </tr>
             </thead>
-
             <tbody>
               {isLoading ? (
                 Array.from({ length: 2 }).map((_, i) => (
@@ -178,7 +125,6 @@ export default function Members() {
                         </div>
                       </div>
                     </td>
-
                     <td className="text-mauve-11 px-4 py-3">
                       {member.is_owner ? "Owner" : "Member"}
                     </td>
@@ -189,7 +135,57 @@ export default function Members() {
           </table>
         </div>
       </div>
-    </div>
+
+      {organization.isOwner && (
+        <Dialog
+          open={showAddMemberDialog}
+          onOpenChange={setShowAddMemberDialog}
+        >
+          <DialogContent>
+            <div className="flex flex-col gap-4">
+              <div className="flex flex-col gap-2">
+                <h1>Invite Member</h1>
+                <p className="text-mauve-11 text-sm">
+                  Invite members via email. They&apos;ll receive a link to join
+                  your organization.
+                </p>
+              </div>
+              <form
+                className="flex flex-col gap-3"
+                onSubmit={handleSubmit(onInviteMember)}
+              >
+                <input
+                  type="email"
+                  className="border-mauve-6 text-mauve-12 placeholder:text-mauve-11 bg-gray-2 w-full rounded-md border p-2 text-sm"
+                  placeholder="Email*"
+                  {...register("email")}
+                />
+                <div className="flex justify-end gap-2">
+                  <Button
+                    type="button"
+                    intent="secondary"
+                    className="w-24"
+                    onClick={() => {
+                      setShowAddMemberDialog(false);
+                      reset();
+                      sendInviteMutation.reset();
+                    }}
+                  >
+                    Cancel
+                  </Button>
+                  <Button
+                    className="h-10 w-24"
+                    type="submit"
+                    disabled={!emailInput || sendInviteMutation.isPending}
+                  >
+                    Invite
+                  </Button>
+                </div>
+              </form>
+            </div>
+          </DialogContent>
+        </Dialog>
+      )}
+    </>
   );
 }
-
